Caddyfile format, attempt to use build time ENV variables
Note there was issue renewing certs with cloudflare plugin. Took a lot
of time to resolve but worked out to be due to either the TLS or
top-level directive not being able to properly read from environment.
Hope this commit fixes it.

Unfortunately have been unable to test renewal since then. Leaving this
as a note in case renewal gets stuck again